+++ internal/provider/provider.go
@@ -24,6 +24,7 @@ import (
 	"github.com/maxvanp/terraform-provider-graviteeam/internal/resources/certificate"
 	"github.com/maxvanp/terraform-provider-graviteeam/internal/resources/deviceidentifier"
 	"github.com/maxvanp/terraform-provider-graviteeam/internal/resources/domain"
+	"github.com/maxvanp/terraform-provider-graviteeam/internal/resources/domainflow"
 	"github.com/maxvanp/terraform-provider-graviteeam/internal/resources/emailtemplate"
 	"github.com/maxvanp/terraform-provider-graviteeam/internal/resources/extensiongrant"
 	"github.com/maxvanp/terraform-provider-graviteeam/internal/resources/factor"
@@ -145,6 +146,7 @@ func (p *GraviteeAMProvider) Resources(_ context.Context) []func() resource.Reso
 		certificate.NewCertificateResource,
 		deviceidentifier.NewDeviceIdentifierResource,
 		domain.NewDomainResource,
+		domainflow.NewDomainFlowResource,
 		emailtemplate.NewEmailTemplateResource,
 		extensiongrant.NewExtensionGrantResource,
 		factor.NewFactorResource,